Proposal for breaking large scenarios out into separate files
Think this might be easier to read and maintain over time.
Wondering whether the schema and data need to be broken out since the
`scenario.yaml` doesn't have much in it any more.

Not sure about number prefixes since this can be tricky if you need to
insert a test in the middle, but it does aid reading by ordering.